+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 211:
 * Improvements to test::stream
 * Add stranded_stream
 * Add flat_stream
+* flat_buffer::clear preserves capacity
 
 --------------------------------------------------------------------------------